+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### BREAKING CHANGES
+- Changed rules from eslint-plugin-unicorn
+- Changed rules from typescript-eslint
+
+### Added
+- Support for eslint-plugin-unicorn@48
 
 ## [11.0.0] - 2023-06-14
 ### BREAKING CHANGES